requirements.txt
attrs==17.4.0
chardet==3.0.4
cookies==2.2.1
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.15
django-log-request-id==1.3.2
django-rest-swagger==2.1.2
djangorestframework==3.7.7
docutils==0.14
pbr==4.0.1
pluggy==0.6.0
psycopg2==2.7.4
py==1.5.3
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2018.3
requests==2.20.0
responses==0.9.0
simplejson==3.13.2
six==1.11.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.17
vine==1.1.4
networkx==2.3